feat(core): split schematics into separate steps
