+++ libs/core/schematics/add-dependencies/schema.json
@@ -1,0 +1,15 @@
+{
+    "$schema": "http://json-schema.org/schema",
+    "$id": "fundamental-ngx-add-dependencies",
+    "title": "Fundamental Library for Angular add-dependencies schematic",
+    "type": "object",
+    "properties": {
+        "project": {
+            "type": "string",
+            "description": "The name of the project.",
+            "$default": {
+                "$source": "projectName"
+            }
+        }
+    }
+}
