+++ libs/core/schematics/add-theming/schema.json
@@ -1,0 +1,15 @@
+{
+    "$schema": "http://json-schema.org/schema",
+    "$id": "fundamental-ngx-add-theming",
+    "title": "Fundamental Library for Angular add-theming schematic",
+    "type": "object",
+    "properties": {
+        "project": {
+            "type": "string",
+            "description": "The name of the project.",
+            "$default": {
+                "$source": "projectName"
+            }
+        }
+    }
+}
