+++ libs/core/schematics/models/schema.ts
