+++ libs/core/schematics/utils/package-utils.ts
@@ -1,7 +1,8 @@
-import { Tree, SchematicsException } from '@angular-devkit/schematics';
+import { Tree, SchematicsException, SchematicContext, Rule } from '@angular-devkit/schematics';
 
 import * as ts from 'typescript';
 import { compare, CompareOperator } from 'compare-versions';
+import { NodePackageInstallTask } from '@angular-devkit/schematics/tasks';
 
 // Gets the ts source file from a path
 export function getSourceFile(host: Tree, path: string): ts.SourceFile {
@@ -45,4 +46,12 @@ export function checkPackageVersion(tree: Tree, name: string, compareTo: string,
     const packageJson = JSON.parse(tree.read('package.json')!.toString('utf-8'));
 
     return compare(packageJson.dependencies[name], compareTo, operator);
+}
+
+export function installDependencies(): Rule {
+    return (tree: Tree, context: SchematicContext) => {
+        context.addTask(new NodePackageInstallTask());
+
+        return tree;
+    };
 }
